fixed redirect error
formerly the styles were not working when inside a subdirectory on the website. Even though it is not possible currently when going to an image /misc and going further in a folder past that, it will not redirect you to the homepage. this has been fixed and the 404 page works as intended for this small error